requirements.txt
pdfminer.six
sentence-transformers
scikit-learn
pandas
plotly
matplotlib
seaborn
numpy
scipy
